requirements.txt
black
coverage==6.2
datasets==1.12.1
flake8
flake8-bugbear
isort
matplotlib==3.5.0
mypy
nltk==3.6.5
numpy
orjson
pytest
pytest-cov==3.0.0
preface==0.1.5
pytorch_lightning
git+https://github.com/samuelstevens/relic.git
scikit-learn
scipy
seaborn
tabulate==0.8.9
tomli==1.2.1
tomli-w==1.0.0
tqdm
transformers==4.11.2
